do not treat as an error if profile file permission is too open
Windows has complicated ACL which may cause issues e.g. our user copies
profile files to USB drive and paste it to another machine, somehow
Administrator & SYSTEM users become accessible to the file.

We stop treating it as an error, just show a warning message and let
users know the profile file is not so secure.