Add --delete-immediately option to "sigfox-devices terminate"
Curerntly Sigfox devices are deleted in twenty four hours after they are
terminated. However, we learned in some cases operators want to delete
them immediately. For instance, the user that has multiple operators
might want to transfer Sigfox devices between those operators. In that
case they might not want to wait for twenty four hours and register
their Sigfox devices for an operator right after terminating the devices
that were owned another operator.

To enable immediate deletion of the Sigfox devices, --delete-immediately
option is added for "sigfox-devices terminate" subcommand. This
immediately deletes the Sigfox device and its PAC is given in the
response. Note it is the responsibility of the operator to keep the PAC
with them.

Signed-off-by: Taku Fukushima <taku@soracom.jp>