avoid generating filenames ends with _test.go
In the Go language, source code whose file name ends in `_test.go` is
treated as test code and is not included in the executable.
soracom-cli uses the rule of concatenating subcommands with underscores
as a file naming convention for auto-generated source code files.

Source code named soralets_test.go is generated for the `soralets test`
subcommand, but this is not included in the executable and the user
could not execute the `soralets test` subcommand.

This commit avoids the issue by adding `_cmd` so that the generated
filename becomes `_test_cmd.go` when it would otherwise be `_test.go`
and is treated as code that should be included in the executable, not
test code.