avoid using backticks in parameter help text
Due to an undocumented specification in Cobra, the library we use to
build our CLI, it turns out that if the help text for a parameter
contains a string enclosed in backticks (`), it is unintentionally used
as the parameter type (or parameter value name). This is almost always
undesirable behavior. For example, the following help text makse no
sense:

```
soracom bills export -- help

...

Flags:
      --export-mode async   Specify how to get the URL to download ...
  -h, --help                help for export
      --yyyy-mm string      Target year and month

...

```

Note that `--export-mode` parameter has `async` as its type. It should
be `string` as same as for `--yyyy-mm`.

This commit avoid this behavior by replacing backticks (`) in the help
text with single quotation marks (').